Fix scipen to prevent errors
When scipen < -1, the year 2000 is stored as scientific notation 2e+3. In climwin, there are cases where year is used as an identifier for cohorts and is converted to character. This can lead to errors as the year 2000 is then stored as '2e+3' and is not longer recognised as a year at later points. During climwin analysis, scipen is now fixed to be >0 so this problem is avoided.