Use simplified update function
Using update to evaluate models is the bottleneck in the slidingwin approach. We can save some time by carrying out the evaluation manually rather than using the update function. The update function causes problems because a) the way it scopes b) it checks for other things (e.g. change of data) that we know are not relevant for our cases. Therefore, by running the reevaluation manually, we can save a small amount of time every time we fit a new model.