Change list appending to pass R 4.0.0
Issue #8

How we append to list objects behaves differently in v4.0.0 compared to previous versions.

**<4.0.0**
```
x <- list()
x$A[[1]] <- 1
x$A[[2]] <- 2
```
Will return a named list with object A containing a vector of numerics. Note that this is the same result as:
```
x <- list()
x$A[1] <- 1
x$A[2] <- 2
```

**>=4.0.0**
```
x <- list()
x$A[[1]] <- 1
x$A[[2]] <- 2
```
Will return a named list with object A containing an unnamed list. Note that this is DIFFERENT behaviour to:
```
x <- list()
x$A[1] <- 1
x$A[2] <- 2
```
Where object A will contain a vector of numerics.

---------------------------

Because of these changes, we need to adjust the code to only use [] when adding items to list objects to ensure compatibility across release and devel version.